FE: move App component to routes/ dir
