+++ frontend/src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 import Root from "./routes/Root.tsx";
 import { RouterProvider, createBrowserRouter } from "react-router-dom";
-import App from "./App.tsx";
+import App from "./routes/App.tsx";
 import "./i18n.ts";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
@@ -27,5 +27,5 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
     <QueryClientProvider client={queryClient}>
       <RouterProvider router={router} />
     </QueryClientProvider>
-  </React.StrictMode>
+  </React.StrictMode>,
 );
